// sentio/morpho-vault/src/schema/schema.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */
import type { String, Int, BigInt, Float, ID, Bytes, Timestamp, Boolean, Int8 } from '@sentio/sdk/store'
import { Entity, Required, One, Many, Column, ListColumn, AbstractEntity, getStore, UpdateValues } from '@sentio/sdk/store'
import { BigDecimal } from '@sentio/bigdecimal'
import { DatabaseSchema } from '@sentio/sdk'

export enum PositionType {
  LP = "LP", GAUGE = "GAUGE"
}





interface AccountConstructorInput {
  id: ID;
  address: String;
  type: PositionType;
  updatedAt: BigInt;
  balance: BigInt;
}
@Entity("Account")
export class Account extends AbstractEntity  {

	@Required
	@Column("ID")
	id: ID

	@Required
	@Column("String")
	address: String

	@Required
	@Column("String")
	type: PositionType

	@Required
	@Column("BigInt")
	updatedAt: BigInt

	@Required
	@Column("BigInt")
	balance: BigInt
  constructor(data: AccountConstructorInput) {super()}
  
  
  static update(values: UpdateValues<AccountConstructorInput>): Promise<void> {
    return getStore().update(Account, values)
  }
}


const source = `type Account @entity {
  id: ID!
  address: String!
  type: PositionType!
  updatedAt: BigInt!
  balance: BigInt!
}

enum PositionType {
  LP
  GAUGE
}
`
DatabaseSchema.register({
  source,
  entities: {
    "Account": Account
  }
})